fix(types): centralise BrowserStrapi type in @strapi/types
