--- packages/admin-test-utils/custom.d.ts
@@ -1,23 +1,0 @@
-export {};
-
-declare global {
-  interface Window {
-    strapi: {
-      backendURL: string;
-      isEE: boolean;
-      features: {
-        SSO: 'sso';
-        isEnabled: (featureName?: string) => boolean;
-      };
-      future: {
-        isEnabled: (name: string) => boolean;
-      };
-      projectType: string;
-      telemetryDisabled: boolean;
-      flags: {
-        nps: boolean;
-        promoteEE: boolean;
-      };
-    };
-  }
-}
