+++ packages/admin-test-utils/tsconfig.build.json
@@ -5,6 +5,6 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src", "custom.d.ts"],
+  "include": ["src"],
   "exclude": ["**/__tests__/**"]
 }
