--- packages/core/admin/admin/custom.d.ts
@@ -1,40 +1,0 @@
-/// <reference types="vite/client" />
-
-import { type StrapiTheme } from '@strapi/design-system';
-
-import type { Modules } from '@strapi/types';
-
-declare module 'styled-components' {
-  export interface DefaultTheme extends StrapiTheme {}
-}
-
-interface BrowserStrapi {
-  backendURL: string;
-  isEE: boolean;
-  future: {
-    isEnabled: (name: keyof NonNullable<Modules.Features.FeaturesConfig['future']>) => boolean;
-  };
-  features: {
-    SSO: 'sso';
-    AUDIT_LOGS: 'audit-logs';
-    REVIEW_WORKFLOWS: 'review-workflows';
-    isEnabled: (featureName?: string) => boolean;
-  };
-  isTrialLicense: boolean;
-  flags: {
-    promoteEE?: boolean;
-    nps?: boolean;
-    docLinks?: boolean;
-  };
-  projectType: 'Community' | 'Growth' | 'Enterprise';
-  telemetryDisabled: boolean;
-  ai: {
-    enabled: boolean;
-  };
-}
-
-declare global {
-  interface Window {
-    strapi: BrowserStrapi;
-  }
-}
