+++ packages/core/admin/admin/src/render.ts
@@ -6,7 +6,7 @@ import { getFetchClient } from './utils/getFetchClient';
 import { getProjectType } from './utils/getProjectType';
 import { createAbsoluteUrl } from './utils/urls';
 
-import type { Modules } from '@strapi/types';
+import type { Admin, Modules } from '@strapi/types';
 
 interface RenderAdminArgs {
   customisations: {
@@ -26,7 +26,7 @@ const renderAdmin = async (
     throw new Error('[@strapi/admin]: Could not find the root element to mount the admin app');
   }
 
-  window.strapi = {
+  const browserStrapi: Admin.BrowserStrapi = {
     /**
      * This ENV variable is passed from the strapi instance, by default no url is set
      * in the config and therefore the instance returns you an empty string so URLs are relative.
@@ -42,8 +42,6 @@ const renderAdmin = async (
         return features?.future?.[name] === true;
       },
     },
-    // eslint-disable-next-line
-    // @ts-ignore – there's pollution from the global scope of Node. Cannot use @ts-expect-error because of build:code and build:types context collision. Cannot use @ts-expect-error because of build:code and build:types context collision.
     features: {
       SSO: 'sso',
       AUDIT_LOGS: 'audit-logs',
@@ -60,8 +58,6 @@ const renderAdmin = async (
       promoteEE: true,
       docLinks: true,
     },
-    // eslint-disable-next-line
-    // @ts-ignore – there's pollution from the global scope of Node. Cannot use @ts-expect-error because of build:code and build:types context collision.
     ai: {
       enabled: true,
     },
@@ -69,7 +65,7 @@ const renderAdmin = async (
 
   const { get } = getFetchClient();
 
-  interface ProjectType extends Pick<Window['strapi'], 'flags'> {
+  interface ProjectType extends Pick<Admin.BrowserStrapi, 'flags'> {
     isEE: boolean;
     isTrial: boolean;
     /**
@@ -92,18 +88,16 @@ const renderAdmin = async (
       },
     } = await get<{ data: ProjectType }>('/admin/project-type');
 
-    window.strapi.isEE = isEE;
-    window.strapi.isTrialLicense = isTrial;
-    window.strapi.flags = flags;
-    window.strapi.features = {
-      ...window.strapi.features,
+    browserStrapi.isEE = isEE;
+    browserStrapi.isTrial = isTrial;
+    browserStrapi.flags = flags;
+    browserStrapi.features = {
+      ...browserStrapi.features,
       isEnabled: (featureName: string | undefined) =>
         features.some((feature) => feature.name === featureName),
     };
-    window.strapi.projectType = getProjectType({ isEE, planPriceId });
-    // eslint-disable-next-line
-    // @ts-ignore – there's pollution from the global scope of Node. Cannot use @ts-expect-error because of build:code and build:types context collision.
-    window.strapi.ai = ai;
+    browserStrapi.projectType = getProjectType({ isEE, planPriceId });
+    browserStrapi.ai = ai;
   } catch (err) {
     /**
      * If this fails, we simply don't activate any EE features.
@@ -111,6 +105,9 @@ const renderAdmin = async (
      */
     console.error(err);
   }
+
+  // @ts-expect-error - conflicting global.Strapi with window.BrowserStrapi
+  window.strapi = browserStrapi;
 
   const app = new StrapiApp({
     config: customisations?.config,
