+++ packages/core/admin/admin/tsconfig.build.json
@@ -4,9 +4,9 @@
     "rootDir": "../",
     "outDir": "../dist",
     "noEmit": false,
-    "types": ["@testing-library/jest-dom"]
+    "types": ["@testing-library/jest-dom", "vite/client", "@strapi/types/globals-admin"]
   },
-  "include": ["./src", "./custom.d.ts", "../shared", "../ee/admin", "../package.json", "./tests"],
+  "include": ["./src", "../shared", "../ee/admin", "../package.json", "./tests"],
   "exclude": [
     "**/__mocks__",
     "./src/**/tests",
