+++ packages/core/admin/ee/admin/tsconfig.json
@@ -4,7 +4,7 @@
     "paths": {
       "@tests/*": ["../../admin/tests/*"]
     },
-    "types": ["@testing-library/jest-dom", "jest"]
+    "types": ["@testing-library/jest-dom", "jest", "vite/client", "@strapi/types/globals-admin"]
   },
-  "include": ["./src", "../../shared", "./tests", "./custom.d.ts"]
+  "include": ["./src", "../../shared", "./tests"]
 }
