+++ packages/core/content-manager/admin/src/globals.d.ts
@@ -1,0 +1,17 @@
+import { type BaseEditor } from 'slate';
+import { type HistoryEditor } from 'slate-history';
+import { type ReactEditor } from 'slate-react';
+
+import type { LinkEditor } from './pages/EditView/components/FormInputs/BlocksInput/Blocks/Link';
+import type { Schema } from '@strapi/types';
+
+declare module 'slate' {
+  interface CustomTypes {
+    Editor: Omit<BaseEditor & ReactEditor & HistoryEditor & LinkEditor, 'children'> & {
+      children: Schema.Attribute.BlocksValue;
+    };
+    Element: Schema.Attribute.BlocksNode;
+    Descendant: Schema.Attribute.BlocksInlineNode | Text;
+    Text: Schema.Attribute.BlocksTextNode;
+  }
+}
