+++ packages/core/content-manager/admin/tsconfig.build.json
@@ -5,6 +5,6 @@
     "outDir": "../dist",
     "noEmit": false
   },
-  "include": ["./src", "./custom.d.ts", "../shared"],
+  "include": ["./src", "../shared"],
   "exclude": ["**/__mocks__", "./src/**/tests", "**/*.test.*"]
 }
