--- packages/core/content-releases/admin/custom.d.ts
@@ -1,27 +1,0 @@
-/// <reference types="vite/client" />
-
-export {};
-
-declare global {
-  interface Window {
-    strapi: {
-      backendURL: string;
-      isEE: boolean;
-      features: {
-        SSO: 'sso';
-        AUDIT_LOGS: 'audit-logs';
-        REVIEW_WORKFLOWS: 'review-workflows';
-        isEnabled: (featureName?: string) => boolean;
-      };
-      future: {
-        isEnabled: (name: string) => boolean;
-      };
-      flags: {
-        nps?: boolean;
-        promoteEE?: boolean;
-      };
-      projectType: 'Community' | 'Enterprise';
-      telemetryDisabled: boolean;
-    };
-  }
-}
