--- packages/core/content-releases/admin/module.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="vite/client" />
