+++ packages/core/content-releases/server/src/globals.d.ts
