+++ packages/core/content-releases/server/tsconfig.eslint.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src", "custom.d.ts"]
+  "include": ["src"]
 }
