+++ packages/core/content-releases/server/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "tsconfig/base.json",
-  "include": ["src", "custom.d.ts", "../shared"],
+  "include": ["src", "../shared"],
   "compilerOptions": {
     "types": ["node", "jest"]
   }
