--- packages/core/content-type-builder/admin/custom.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="vite/client" />
-
-import { type StrapiTheme } from '@strapi/design-system';
-
-declare module 'styled-components' {
-  export interface DefaultTheme extends StrapiTheme {}
-}
