+++ packages/core/content-type-builder/admin/tsconfig.build.json
@@ -5,6 +5,6 @@
     "outDir": "../dist",
     "noEmit": false
   },
-  "include": ["./src", "./custom.d.ts"],
+  "include": ["./src"],
   "exclude": ["**/tests/**"]
 }
