+++ packages/core/content-type-builder/admin/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "tsconfig/client.json",
-  "include": ["src", "./custom.d.ts"],
+  "include": ["src"],
   "compilerOptions": {
-    "types": ["@testing-library/jest-dom", "jest"]
+    "types": ["@testing-library/jest-dom", "jest", "vite/client"]
   }
 }
