+++ packages/core/core/src/globals.d.ts
