+++ packages/core/database/src/globals.d.ts
