+++ packages/core/permissions/src/globals.d.ts
