+++ packages/core/review-workflows/server/tsconfig.build.json
@@ -5,6 +5,6 @@
     "outDir": "../dist",
     "noEmit": false
   },
-  "include": ["./src", "./custom.d.ts", "../shared"],
+  "include": ["./src", "../shared"],
   "exclude": ["**/*.test.ts"]
 }
