+++ packages/core/strapi/src/globals.d.ts
