+++ packages/core/types/src/admin/browser-strapi.ts
@@ -1,0 +1,26 @@
+import type { Features } from '../modules';
+
+export interface BrowserStrapi {
+  backendURL: string;
+  isEE: boolean;
+  isTrial: boolean;
+  future: {
+    isEnabled: (name: keyof NonNullable<Features.FeaturesConfig['future']>) => boolean;
+  };
+  features: {
+    SSO: 'sso';
+    AUDIT_LOGS: 'audit-logs';
+    REVIEW_WORKFLOWS: 'review-workflows';
+    isEnabled: (featureName?: string) => boolean;
+  };
+  flags: {
+    promoteEE?: boolean;
+    nps?: boolean;
+    docLinks?: boolean;
+  };
+  projectType: 'Community' | 'Growth' | 'Enterprise';
+  telemetryDisabled: boolean;
+  ai: {
+    enabled: boolean;
+  };
+}
