+++ packages/core/types/src/admin/index.ts
@@ -1,0 +1,1 @@
+export type { BrowserStrapi } from './browser-strapi';
