+++ packages/core/types/src/globals-admin.ts
@@ -1,0 +1,7 @@
+import type { BrowserStrapi } from './admin';
+
+declare global {
+  interface Window {
+    strapi: BrowserStrapi;
+  }
+}
