--- packages/core/upload/admin/custom.d.ts
@@ -1,13 +1,0 @@
-export {};
-
-declare global {
-  interface Window {
-    strapi: {
-      backendURL: string;
-      future: {
-        isEnabled: (name: keyof NonNullable<Modules.Features.FeaturesConfig['future']>) => boolean;
-      };
-    };
-  }
-  module '*?raw';
-}
