+++ packages/core/upload/admin/tsconfig.json
@@ -4,7 +4,7 @@
     "paths": {
       "@tests/*": ["./tests/*"]
     },
-    "types": ["@testing-library/jest-dom", "jest"]
+    "types": ["@testing-library/jest-dom", "jest", "vite/client", "@strapi/types/globals-admin"]
   },
-  "include": ["../package.json", "./src", "../shared", "./tests", "./custom.d.ts"]
+  "include": ["../package.json", "./src", "../shared", "./tests"]
 }
