+++ packages/core/utils/src/globals.d.ts
