+++ packages/generators/generators/src/globals.d.ts
