--- packages/plugins/cloud/admin/global.d.ts
@@ -1,9 +1,0 @@
-export {};
-
-declare global {
-  interface Window {
-    strapi: {
-      backendURL: string;
-    };
-  }
-}
