--- packages/plugins/cloud/admin/module.d.ts
@@ -1,2 +1,0 @@
-declare module '*.svg';
-declare module '*.png';
