--- packages/plugins/documentation/admin/src/global.d.ts
@@ -1,25 +1,0 @@
-export {};
-
-declare global {
-  interface Window {
-    strapi: {
-      backendURL: string;
-      isEE: boolean;
-      features: {
-        SSO: 'sso';
-        AUDIT_LOGS: 'audit-logs';
-        REVIEW_WORKFLOWS: 'review-workflows';
-        isEnabled: (featureName?: string) => boolean;
-      };
-      future: {
-        isEnabled: (name: string) => boolean;
-      };
-      flags: {
-        nps?: boolean;
-        promoteEE?: boolean;
-      };
-      projectType: 'Community' | 'Enterprise';
-      telemetryDisabled: boolean;
-    };
-  }
-}
