+++ packages/plugins/documentation/server/src/globals.d.ts
@@ -1,3 +1,4 @@
+// https://github.com/bdadam/rollup-plugin-html
 declare module '*.html' {
   const content: string;
   export default content;
